Update backend implementation

+++ employee-manager-backend/src/main/java/com/example/employeemanager/controller/EmployeeController.java
@@ -1,7 +1,11 @@
 package com.example.employeemanager.controller;
 
-import com.example.employeemanager.model.Employee;
+import com.example.employeemanager.data.dto.EmployeeDto;
+import com.example.employeemanager.data.dto.EmployeeIncomingDto;
+import com.example.employeemanager.data.model.Employee;
 import com.example.employeemanager.service.EmployeeService;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,44 +18,76 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/v1/employees")
 public class EmployeeController {
+
+    private final ModelMapper modelMapper;
+
     private final EmployeeService employeeService;
 
     @Autowired
-    public EmployeeController(EmployeeService employeeService) {
+    public EmployeeController(EmployeeService employeeService, ModelMapper modelMapper) {
         this.employeeService = employeeService;
+        this.modelMapper = modelMapper;
     }
 
     @GetMapping
-    public ResponseEntity<List<Employee>> getAllEmployees() {
+    public ResponseEntity<List<EmployeeDto>> getAllEmployees() {
         List<Employee> employees = employeeService.findAllEmployees();
-        return new ResponseEntity<>(employees, HttpStatus.OK);
+        if (employees.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        List<EmployeeDto> list = employees
+                .stream()
+                .map(employee -> modelMapper.map(employee, EmployeeDto.class))
+                .collect(Collectors.toList());
+        return new ResponseEntity<>(list, HttpStatus.OK);
     }
 
     @GetMapping("{employeeId}")
-    public ResponseEntity<Employee> getEmployee(@PathVariable("employeeId") Long id) {
-        Employee employee = employeeService.findEmployee(id);
-        return new ResponseEntity<>(employee, HttpStatus.OK);
+    public ResponseEntity<EmployeeDto> getEmployee(@PathVariable("employeeId") Long id) {
+        Optional<Employee> employee = employeeService.findEmployee(id);
+        if (employee.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        EmployeeDto employeeDto = modelMapper.map(employee, EmployeeDto.class);
+        return new ResponseEntity<>(employeeDto, HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<Employee> addEmployee(@RequestBody Employee employee) {
+    public ResponseEntity<EmployeeDto> addEmployee(@Valid @RequestBody EmployeeIncomingDto employeeIncomingDto) {
+        Employee employee = modelMapper.map(employeeIncomingDto, Employee.class);
         Employee addedEmployee = employeeService.addEmployee(employee);
-        return new ResponseEntity<>(addedEmployee, HttpStatus.CREATED);
+        EmployeeDto employeeDto = modelMapper.map(addedEmployee, EmployeeDto.class);
+        return new ResponseEntity<>(employeeDto, HttpStatus.CREATED);
     }
 
     @PutMapping("{employeeId}")
-    public ResponseEntity<Employee> updateEmployee(@PathVariable("employeeId") Long id, @RequestBody Employee employee) {
-        Employee updatedEmployee = employeeService.updateEmployee(employee);
-        return new ResponseEntity<>(updatedEmployee, HttpStatus.OK);
+    public ResponseEntity<EmployeeDto> updateEmployee(
+            @PathVariable("employeeId") Long id,
+            @Valid @RequestBody EmployeeIncomingDto employeeIncomingDto) {
+        Optional<Employee> optionalEmployee = employeeService.findEmployee(id);
+        if (optionalEmployee.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        BeanUtils.copyProperties(employeeIncomingDto, optionalEmployee.get());
+        Employee updatedEmployee = employeeService.updateEmployee(optionalEmployee.get());
+        EmployeeDto employeeDto = modelMapper.map(updatedEmployee, EmployeeDto.class);
+        return new ResponseEntity<>(employeeDto, HttpStatus.OK);
     }
 
     @DeleteMapping("{employeeId}")
     public ResponseEntity<?> updateEmployee(@PathVariable("employeeId") Long id) {
+        Optional<Employee> employee = employeeService.findEmployee(id);
+        if (employee.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         employeeService.deleteEmployee(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
